Remove useless invariant in gc_alloc_large()
The FIXME comment at 'gc_assert(first_page > alloc_region->last_page)'
is right - it is a nonsense assertion. The previous change did not offer
a good explanation as to why though.

Suppose the largest contiguous free block below last_free_page is 4 pages.
This the minimal "large object" allocation.
On 32-bit implementations, allocating a non-large code component rounds the
allocation size up to 64K (16 pages) to reduce fragmentation. This is larger
than 4 pages, so to allocate this size, we extend the heap's frontier and bump
up last_free_page to coicide with the code region's end page. Now suppose we
want to make a 4-page "large" object - it fits in the gap that is somewhere
below last_free_page. But for no good reason, we asserted that it goes *after*
the small object region. The purpose of aways scanning forwards is to reduce
work, but it's one thing to have that heuristic, and another to say that it
always applies when it does not.  Specifically, in the final GC prior to save,
we set 'gencgc_alloc_start_page' to control the start of scanning which forces
every linear scan to start at that page. (This itself is not efficient, which
is irrelevant) However, it is obviously incorrect to assert that allocations
begin on strictly monotonically increasing page numbers.