debug-int: Extract PC-OFFSET calculation from non-x86oid FIND-ESCAPED-FRAME
  * It's a fairly compact chunk of code, duplicated in a couple of
places, and could be useful more generally as a separate function.

  * One bit of cleverness here is that we pin the CODE object
while we mess with the program counter.  The program counter is an
interior pointer, and is either bounded by the CODE object (and
thus won't move) or NOT bounded by the CODE object (and thus if it
DOES move it won't move into the bounds of the CODE object).  As
such, this new function does not need to be called from within a
WITHOUT-GCING.