ir2tran: COPY-MORE-ARG doing XEP-SETUP-SP's job is arch-specific
  * COPY-MORE-ARG doing XEP-SETUP-SP is arch-specific, having
nothing to do with precise-arg-count-error.  The case that this
was originally trying to protect against was taking an error trap
before the stack frame was sufficiently initialized to do a
backtrace, but that still happened on precisely the systems
supposedly being protected, and has since been fixed.

  * So, rather than use a misleading tangentially-related feature
conditional, tighten up the logic generally and test for affected
architectures.