pack: If we need to spill to a specified-save TN, allocate there to begin with
  * This is, impressively enough, a debugger issue.  There are
potential performance impacts as well, but they are substantially
more general, and the correctness issue is specific to TNs with
specified save locations (the old-fp and return-pc save locations).

  * In principle, knowing the PC offset within a code object
should be sufficient to determine where to look for every bit of
data relevant to the current computation.  In practice, the
compiler only produces accurate maps of where things are at
specific offsets (code locations) within the function.  This is
fine when there's anything resembling a guarantee that these
locations are where the debugger stops.

  * This is all well and good for the program variables and such,
but it also applies to the return-pc and old-fp (LRA and OCFP),
the two parameters which let the debugger find the previous stack
frame.  Because asynchronous interrupts (and even error traps!)
are a thing, these two pieces of data must be findable no matter
where in the function execution stops.

  * It is acceptable for these values to be stored in registers,
provided that they do not get spilled to the stack.  This scenario
happens with "leaf functions" that call no other functions (all of
their operations are open-coded by the compiler).

  * If one of these values gets stored in a register and a full
call occurs, it will end up getting spilled to the stack.  The
function writes it from the register to the designated save slot
(fine), and calls out (fine).  If the debugger does a backtrace
here it sees that this stack frame "is not an escaped frame", i.e.
there is no interrupt context, and looks on the stack for the OCFP
and LRA (this is a kludge).  The PC-offset here is the point
immediately following the LRA header.  If, when the callee returns
to this point, an interrupt occurs, and the debugger tries to
backtrace again, it sees that there is an interrupt context, and
uses the function's debug-info to see where the OCFP and LRA are
stored: in a register that hasn't actually been reloaded yet (not
good, the debug-info is wrong at this point).

  * So, that's the problem: the debug-info is wrong.  We have two
(well, three) options: we can change the debug-info to be right,
by tracking spills and restores and creating a map of where the
value is at all times, a fairly involved endeavour (and one that
we may attempt another time, for other reasons); we can change
the debug-info to be right, by having the compiler produce code
that matches what the debug-info claims (what we're doing here);
or we can ignore the problem as unimportant (words fail me).

  * This choice forces a particular code-generation strategy,
which may in fact be non-optimal in some cases.  It is possible
to set up a scenario in which there is a full-call, forcing the
spill, and then a multi-way control-flow split in which each path
has to reload the spilled value independently, while having the
values stored in registers would cause the reload to happen
before the control-flow split.  On the other hand, it is also
possible to set up such a scenario where there are multiple full
calls, where the value being packed to a register causes it to be
reloaded multiple times, only to be discarded without being used.
I do not expect sorting that mess out to be trivial, and it is
certainly out-of-scope for me this month.

  * Oh, and all this only applies to systems using specified-save
TNs, which means #!-fp-and-pc-standard-save, so most of you who
are (or aren't!) reading this commit message won't see any direct
benefit from this change.