debug-int: Don't randomly corrupt the heap in MAKE-BOGUS-LRA
  * What looked to be a form trying to set up the LRA header was,
in fact, writing a VALUE-CELL-WIDETAG with no header data to the
object two words before the allocated code object, at least on PPC.

  * Did not affect x86oids, of course, because they follow
different rules.  And not detectable on ARMoids because they don't
even have the fun_end_breakpoint_guts defined.

  * And the code doesn't appear to be necessary, because the LRA
header is STILL part of fun_end_breakpoint_guts everywhere that
it's both implemented and necessary.  As such, kill it.