debug-int: Make (SETF FRAME-SAVED-LRA) and use it from FUN-END-STARTER-HOOK
  * FRAME-SAVED-LRA is more accurate than anything involving
GET-CONTEXT-VALUE, because it accounts for details like the LRA
possibly still being in its passing location.

  * Make (SETF FRAME-SAVED-LRA) so that we can use it instead of
(SETF GET-CONTEXT-VALUE), and update FUN-END-STARTER-HOOK to use
these.

  * This causes the (TRACE :ENCAPSULATE NIL :RECURSIVE) test to
stop failing on PPC/Linux...  But I remain unconvinced as to the
fundamental soundness of the mechanism (compiled breakpoints, as
used by TRACE :ENCAPSULATE NIL, are currently unsafe).