Make an 'if' test in update_page_write_prot() straightforward
The first series of tests for setting 'wp = 0' was disingenuous and suboptimal.

The region-based criteria, as used, were bad for 2 reasons:
* At first glance, we falsely imply that the generation of the pointed-to page
  is irrelevant, due to testing the generation only if bytes_used were nonzero;
  and pages with open regions usually, but not always, have 0 bytes in use.
* Regions bounds were *always* tested when the generation test fails.

As to the first point, an open region must be for the same, a lower, or the temp
generation. But it's right by such a subtlety that it looks wrong - how can
we know that it's a strictly lower generation when it isn't the temp generation?
Because root pages for scavenge_generations() do not include newspace. Namely:
 * on a non-raise cycle (gen N is copied to 7 and moved en masse back to N),
   the roots are N+1 through 6;
 * on a 'raise' cycle (gen N is copied to N+1), the roots are N+2 through 6,
   and we assume that all of N+1 is live. It is scavenged as part of newspace,
   and not the root set.

So the improvement here is that we should only test regions if the page
generation is actually lower - we shouldn't *always* test. And moreover,
only test if the bytes_used test fails *and* the page has an open region.
So expressing this 'if' as a generation criterion and a page-in-use criterion
is faster as well as far more obviously correct.