ir2tran: Fix commentary and conditionals for COPY-MORE-ARG damage
  * When last I was adjusting this logic, the fact that there was
an interrupt-safety hole being covered here had escaped me.  And
because it's an interrupt-safety thing, we need to transition all
backends to the newer form of COPY-MORE-ARG, and that means having
a list of non-converted backends, not a list of converted backends.

  * Note that we usually transition from a list-of-fixed-backends
to a list-of-still-broken-backends at about the halfway point.
We're at 4/9 fixed now, and should be at 5/9 soon.

  * Also fix compiler.pure.lisp / (COMPILE :COPY-MORE-ARG)
:FAILS-ON in the same way: It's a correctness bug, it has been
fixed for the exact same set of architectures, and we may expect
that both issues will be fixed at the same time, since they are
handled by the same VOP.