PPC COPY-MORE-ARG fix 4/4: Copy more-args in an appropriate direction
  * The direction in which the more-args need to be copied depends
on the relation between the size of the fixed stack frame and the
number of fixed arguments.  In some cases, we don't need to copy
at all.

  * Someone might be able to come up with a tighter code sequence
for the "negative DELTA" case, but this works well enough for now.

  * And the little shortcut that prevents copying extra data for
the arguments that are passed in registers got disabled in this
change.  It doesn't cause incorrect operation or new GC hazards,
just unnecessarily copying up to register-arg-count additional
args.