Reinstate removal of 'has_pins' bit, with bug fix
There was a small difference in behavior of 'has_pins' versus the supposed
equivalent expression (dont_move && !large_object).

The difference is that dont_move bits are cleared lazily, and not zeroed
prior to full mark-and-sweep GC, whereas 'has_pins' was always initially zero
and after wipe_nonpinned_words(), there can be no pages with 'has_pins'.
Using dont_move as the replacement for has_pins, there were pages marked
dont_move while doing pointer preservation, so pinned_p() might spuriously
be true. The purpose of the pinned_p() check there is to avoid heap scan more
than once for the same pointer appearing multiple times on the stack(s).

The simple fix is to always clear the pinned_objects hash table whether it
will be used or not. It might also make sense to use the hash-table
during mark-and-sweep, though it is strictly redundant. Te debug check in
pinned_p - that would have caught this sooner - would need to be removed.