Remove "start" pages from write_generation_stats()
Given that we can only allocate to one generation (gen0 from Lisp, or newspace
during GC), and reset all start pages to 0 before and after every GC,
the "starting page for allocation" is well near meaningless. I plan to remove
the per-generation tracking of those, so might as well remove the display first,
as it makes room for wider columns. It's high time those were enlarged.