Allow per-file granularity for !unintern-init-only-stuff
Rather than define a function per package to remove unneeded symbols,
build up a list of symbols that can be removed. This is updatable many
times per package so that one file is no longer the designated driver
for a given package (such as SB-IMPL). Keep the package-specific cleanup
function working as well since the disassembler uses it it.