Remove an AVER from VALIDATE-CALL-TYPE.
It was trying to ascertain that the type is indeed for a function, but
it doesn't accept union types. Instead of trying to use an
intersection just don't assert anything, if the compiler is so broken
to get to this point with a non-function type it'll become evident at
some later stage.

Fixes lp#1748035