Lift cross-compiler limitation on use of non-EQL-comparable constants
DEFCONSTANT-EQX was exceedingly tricky if not impossible to use in cross-
compiled code. This stems from its ability to define constants whose value form
is not constant per se, thus not evaluable at compile-time (a constructor call
for example), and so deferred until cold init; whilst each use of the symbol
so defined was emitted as (LOAD-TIME-VALUE (SYMBOL-GLOBAL-VALUE 'CONST)),
assumed to be evaluable during fasl loading. But the cold loader can't execute
target machine code, and so mightn't have set a value for the constant at all.

Thus it only "sometimes" worked to use DEFCONSTANT-EQX. This change makes
it reliable. If the value form does not satisfy CONSTANTP, then any use of
the constant is *not* looked up at cold load time. It reverts to the behavior
of compiling the LTV form into target machine code for cold-init,
and then deferring the pertinent fixups until cold-init too.