Deal with imprecision of or misinterpretation of Common Lisp spec
(I'm looking at you, Clozure Common Lisp)

Prior changes caused a problem with non-EQL-comparable constants having naught
to do with redefinition thereof (undefined behavior), but definition at all!

Apparently the compiler can force / require you to define a MAKE-LOAD-FORM
method on objects that did not appear as literals in your code.

In particular, the following minimal example does not work in CCL:
== /tmp/example.lisp ==
(eval-when (:compile-toplevel :load-toplevel :execute)
  (defstruct thing))
(defconstant athing (make-thing))
(defun somefun () (list athing))
== EOF ==
% Welcome to Clozure Common Lisp Version 1.10-r16196  (LinuxX8664)!
...
? (COMPILE-FILE "/tmp/example.lisp")
> Error: No MAKE-LOAD-FORM method is defined for #S(THING)

The issue seems to be that substitution of the constant occurs early,
i.e. the DEFUN body effectively becomes (list #S(THING))
and so it looks like it has unquoted data in it.  Ref CLHS:
* literal adj. (of an object) referenced directly in a program rather than
being computed by the program; that is, appearing as data in a quote form,
or, if the object is a self-evaluating object, appearing as unquoted data.