Record relative fixups from immobile code to linkage table addresses
This is necessary due to change 85492cc224fd which made calls to linkage
table entries not consume a register. Immobile space defrag handled that fine,
but heap relocation on startup didn't, so we need to store the address of the
CALL operand, and note that the value is PC-relative.

Relative fixups account for typically less than 1/1000th of all fixups,
so the storage technique employed to distinguish absolute from relative
fixups is designed to avoids a size increase if there are none of the latter.

Also some improvements to the mmap mock which would be nice to enable
as part of the test suite:
 * allow the same file of mock data to work with 32-bit or 64-bit build
 * 32-bit x86 linux still needs mmap64() as the entry point