Fix C memory leak in APROF-START
Don't create a new profile buffer if it's same size as the old,
and deallocate the old if changing size.

The only way this causes a crash-inducing race is the following:
 - profiling is started
 - a thread has just performed a load of the bufffer address to a register
 - thread is interrupted and runs other code that takes a while
 - profiling is stopped and restarted with new buffer
 - interrupted thread resumes and stores through the old buffer pointer

Given that the API uses a good default size, reallocation should not happen.
Therefore the above doesn't occur, making this defect not worth fixing.
Incidentally, it is bad idea to shrink the buffer, so I might want to
make the interface refuse to do that.