sb-sprof: Do not record implausibly short traces
It occasionally happens that SIGPROF-HANDLER runs in a situation in
which it cannot walk the stack beyond the initial frame.

Recording such degenerate traces would disrupt certain report kinds
such as flamegraphs. In addition, this seems to happen rarely enough
that just dropping the traces shouldn't really affect statistics-based
reports.