Reduce COMPILED-DEBUG-INFO size by 25% if compact instance header
Shrinks from 8 to 6 words; unchanged if no compact headers.

In my Lisp core which has > 188MB of application-related structures in a
freshly built image, COMPILED-DEBUG-INFO was still a big enough space hog
that it appeared in the top 15 consumers as per SB-VM::INSTANCE-USAGE.
This change moves it down one notch in the ranking.