Fix storage of nested source forms from COMPILE (or LOAD)
This change adds the ability for each simple-fun in a code component
to store its own source form.

In the old way, a code component which contained multiple simple-funs would
store a single form in the code-debug-info of the code component.
A chosen representative function (the toplevel) from the component was stored
as DEBUG-SOURCE-FUNCTION so that we could remember whose source it was,
if more than one visible function. (Optional dispatches and XEP aren't visible).

So when more than one function existed, to inquire whether a function had a
source form, we checked the debug-source of the debug-info and asked whether
the function being queried was specifically the chosen function whose form
was stored. If not, we fell back to namestring. (Logic in GET-TOPLEVEL-FORM)

Now COMPUTE-ENTRY-INFO gets the correct source form for every simple-fun
so that we can attach to that simple-fun. The form goes in %SIMPLE-FUN-INFO
along with what was already there. The accessors for form/doc/xrefs figure
out which was saved. (Because 6 words is a nice even number of words per
function header, and 7 is not as nice, thus the design goal of no new slots)

We could eventually advertise in NEWS that FUNCTION-LAMBDA-EXPRESSION
works better, but not yet - the (DEBUGGER SOURCE 2) test in debug.impure.lisp
fails; and maybe some other undiscovered rough edges still exist.