Implement automatic dxification of funargs to higher-order functions
(https://bugs.launchpad.net/sbcl/+bug/383078)

As per the CLHS writeup of DYNAMIC-EXTENT:
"It is permissible, but in practice probably not as useful, to write:
 (defun g (x) (declare (dynamic-extent x)) ...)
 (defun f () (g (list 1 2 3)))"

So now we will take advantage of that permission and emit stack-allocation
code for objects X of type FUNCTION, but only functions and nothing else;
and only in unsafe code or where the higher-order function is a standard
function. (Which we might have handled via transform already such as MAPC)

I might relax the "only in unsafe code" restriction, but the same users
who want this have expressed concern that this transform is harmful.
It's not per se, except when you redefine the HOF so that a funarg that was
formerly downward becomes upward. If you do that, a warning is issued,
whether or not any unsafe code was compiled.