Pull interrupt_contexts[] out of 'struct thread'
For all backends except x86-64, the array still physically follows the fixed
part of struct thread. For x86-64, the array precedes, which is better because:

  - it frees up 1024 more indices for Lisp's use
  - scavenges less in GC (at the "do the tls as well" comment)
  - generally makes more sense than having an overlap

Additionally, on x86-64, we could store a few important values in the 16 words
before 'struct thread' which are accessible using a 1-byte EA displacement.
(Simply avoid using the first few negatively indexed interrupt context words
for interrupt contexts. There's a new C macro that can readily do that)

Work on other backends to follow. CURRENT-THREAD-OFFSET-SAP needs to
be changed to allow a signed offset, that's all.