Remove weaselly behavior of cross-compiler SXHASH transforms
Be up front about whether the transform wanted to run.

And there's a perfectly good reason, but not yet - if we ever implement [E]CASE
over symbols as a jump table based on a minimal perfect hash of the precomputed
SYMBOL-HASHes (which should be easy: just find some distinctive bits across all
the hashes of the symbols in question), then we would want to know if
this optimization was missed.

Also move the file to where it belongs as part of the compiler,
and delete obsolete comment about dependence on DEFINE-MODIFY-MACRO.