Trace assembly instructions earlier
All machine encoding is done into a single segment, so it doesn't make sense
for the assembler to print segment-type any more. Instead, distinguish sections
when accumulating instructions in symbolic form prior to machine-encoding.
This approach changes how a trace file looks with the scheduling assembler,
but (I think) the purpose of the first trace is to show what instructions
get selected by each vop, and not the effect of the assembler's scheduler.
DISASSEMBLE-ASSEM-SEGMENT shows the latter.