Lazily change linux personality, or not at all
No longer preemptively call personality() always.  Wait to see if our fixed
spaces can be allocated where mandated. If not, change personality and retry.
In effect, it should almost never be necessary to change personality.
However, with a 32-bit address space, it is entirely likely that dynamic space
will have to be relocated on any given invocation. It is of course random.