Remove :immobile-code-pc as a register size
In order to support PIE executables (or even just placement of both immobile
subspaces in a range straddling the 4GiB boundary), some small GC changes
will be needed, but nothing fundamental. The only hard constraints are that:
 - layout pointers fit in the high half of instance header words
 - the distance from fdefns to functions is (SIGNED-BYTE 32)

So we shouldn't wire in more restrictive assumptions to Lisp's codegen.

[The first constraint can even be removed by treating the instance header's
layout as an index to a table of all layouts, placed anywhere.
It would affect type-checking, but not as much as first appears,
because table indices can be directly compared to test whether an object is
of a given type, before using the fallback code that needs LAYOUT-INHERITS,
which would of course entail getting a layout given its index]