Fix regression due to TOP-LEVEL-NAMED-LAMBDA by removing it
The concept remains, but the previous implementation caused the compiler to trip
over its own shoelaces while trying to walk (pun intended), and caused code
analyzers to go awry.  It's near impossible to produce a minimal reduction
because by definition all examples seem to involve a lot of hair: macros that
expand inside-out by MACROEXPANDing their arguments, and somehow accidentally
misled the compiler to thinking that the inside code was toplevel.

The failures in previously working code were several. One was an attempt by
the compiler to call GET-DEFINED-FUN on macrolets (redacted backtrace below),
which causes an error in source-path finding, which we might want to fix as a
bug in its own right, except that it goes away after this patch.

Attempting to hack around that in IR1-CONVERT-LAMBDALIKE by simply never calling
GET-DEFINED-FUN on things named (MACROLET x) produces a different error, which
is the attempt to store (MACROLET mumble) into a %DEBUG-NAME of a CLAMBDA:
 'The value (MACROLET mumble) is not of type (OR NULL (NOT (SATISFIES SB-INT:LEGAL-FUN-NAME-P)))
 when setting slot SB-C::%DEBUG-NAME of structure SB-C::CLAMBDA'

So this changes removes MACROLET as a legal fun name, which makes it never reach
the named lambda case in IR1-CONVERT-LAMBDALIKE, and moreover, MAKE-MACRO-LAMBDA
never returns a top-level-ish lambda from a type derivation perspective.
I can't explain why these never mattered, because it sure looks as if named
macro lambdas always went through the named-lambda path that global functions
did. But at exactly the change which introduced TOP-LEVEL-NAMED-LAMBDA,
without this patch, we would see in various guises:

 SB-INT:INVALID-ARRAY-INDEX-ERROR: Invalid index 41 for (SIMPLE-VECTOR 1), should be
 a non-negative integer below 1. while executing: COMPILE
Unhandled SB-INT:INVALID-ARRAY-INDEX-ERROR in thread #<SB-THREAD:THREAD "main thread" RUNNING {10006F05B3}>:
Invalid index 41 for (SIMPLE-VECTOR 1), should be a non-negative integer below 1.
Backtrace for: #<SB-THREAD:THREAD "main thread" RUNNING {10006F05B3}>
0: (SB-C::FIND-SOURCE-ROOT 41 #<SB-C::SOURCE-INFO {100F483203}>)
1: (SB-C::FIND-ORIGINAL-SOURCE (#1=(SB-C::%FUNCALL #2=#:G23 . #3=(#:G22))
    #4=(LET ((#5=#:VAL26 #1#)) (UNLESS #5# (RETURN-FROM #6=#:BLOCK24 NIL)))
         (BLOCK #7=#:WRAPPER25 . #8=(#4#)) ...))
2: (SB-C::FIND-ERROR-CONTEXT ((MACROLET DEFINE-mumble)) NIL)
3: (SB-C::NOTE-UNDEFINED-REFERENCE (MACROLET DEFINE-mumble) :FUNCTION)
4: (SB-C::FIND-GLOBAL-FUN (MACROLET DEFINE-mumble) NIL)
5: (SB-C::FIND-FREE-FUN (MACROLET DEFINE-mumble) "shouldn't happen! (defined-fun)")
6: (SB-C::GET-DEFINED-FUN (MACROLET DEFINE-mumble) (#:EXPR #:ENV))
7: (SB-C::IR1-CONVERT-LAMBDALIKE (SB-INT:NAMED-LAMBDA (MACROLET DEFINE-mumble)
     (#1=#:EXPR #2=#:ENV) (DECLARE (IGNORE #2#))
       (DECLARE (OPTIMIZE (SB-EXT:INHIBIT-WARNINGS 1))) ...))