Fix altstack_pointer_p()
This is what happens when you have 12 different implementations of the same
thing (which I nearly fixed in da2690ca62) - all of them are wrong.

nonpointer data are not part of the stack; and its high bound was too low
since calc_altstack_size() uses the address of the actual end of the mapped
region which can be farther than 32*SIGSTKSZ away from the 'struct thread'
depending on whether the memory we get from the OS was already aligned enough
for our purposes. (i.e. if it was aligned, then we didn't bump up the base
address the OS give us, but we had already allowed for THREAD_ALIGNMENT_BYTES
of slack bytes which is considered part of the altstack)