x86-64: avoid use of mprotect() on pages containing code
This is an extremely restricted implementation of "manual" (in contrast
to signal-based) card marking designed only to deal with code pages.

Among the benefits are:
 - permitting (in theory) some pages to be mapped to memory with MAP_HUGETLB
   by removing the constraint of protecting at a granularity that the OS wants
   *and* that is a good card size- cards can be as small as desired, but backed
   by 2MB or 1GB pages in harware. This is mainly pertinent to text-in-ELF.

 - implementing an improved variant of update_page_write_prot() which assumes
   that almost all words on a page are non-pointers, as opposed to the general
   case which assumes that all words are pointers.

 - allowing use of the unboxed code constant area as working storage for
   runtime instrumentation without triggering the page write barrier.

Rather than getting too fancy about the precise mechanism by which card marks
are toggled atomically with respect to any given store operation (as they must
be to preserve the marking invariant), this scheme uses pseudo-atomic as
a blunt tool which is perfectly fine given that code objects are almost never
mutated after creation. A better technique will be needed when implementing this
for all pages, as pseudo-atomic has unacceptable overhead for every store.