x86: use new coverage mark representation
Touch up comments, and change main to refer to sb!xc:*compile-file-pathname*
in case we wanted to instrument cross-compilation (which doesn't work).
I didn't benchmark the 32-bit instrumentation, but it certainly seems
that the same rationale as to improved efficiency of 64-bit pertains here,
especially the elision of redundant marks.

Incidentally there's another benefit of the single-bit mark: despite using
a symbol in SB-C as the mark (in the old representation), it's still possible
for COMPILE-FILE to coaelsce any user-written literal that is EQUAL to say
((4 1 0) . %CODE-COVERAGE-UNMARKED%). We'd suprise the user by changing that
constant at runtime. This is why we used an "obscure" symbol.