Speed up PACK-ITERATIVE.
Compute COLOR-NO-CONFLICTS-P in constant time with a simple LOGTEST.
Previously it iterated over all neighbors and checked that the color
+ size intervals do not intersect, which is quite slow.
Compute a mask in COLOR-VERTEX instead.
One slight complication: multiple neighbors can share the color and
the color of a vertex can change, so also compute an array with color
counts, that way it can be used to correctly recompute the mask upon
recoloring.

This makes building Ironclad almost twice as fast.