gencgc: Always collect through gen5 on save
This is due to a new problem caused by manual card marking of code cards.
Similar object coalescing can replace arbitrary words, and assumes that doing
so never creates an intergenerational pointer.  That will be true if the entire
heap is generation 5, as it normally would be at that point, but might not be
if the prior "full" collection stopped short of HIGHEST_NORMAL_GENERATION.
Combine that with the fact that the C code never manually marks cards,
and you could end up with invariant violations after the pre-final GC.