Fix EOF detection in WAIT-UNTIL-FD-USABLE.
With sockets, EOF is not properly detected because WSAIoctl with
FIONREAD does not report input on EOF (there are no bytes to read, after
all). This is fixed by replacing WSAIoctl with select(), which is able
to detect EOF and report "input" available for a subsequent read.

Waiting on a named pipe has a related problem, where PeekNamedPipe
reports EOF as an error, but SB-WIN32:HANDLE-LISTEN does not distinguish
between EOF and a regular read error, and the listen falls through into
the socket code. This produces a functionally identical result, because
errors are treated as input-available, so the resulting not-a-socket
error causes HANDLE-LISTEN to correctly return T. This is surprising,
and ties a pipe listen to the socket code, so instead we now explicitly
check for EOF in the pipe code instead of falling through.

This also adds a test that checks 1) that SB-SYS:WAIT-FOR-INPUT doesn't
hang, and 2) that LISTEN correctly returns NIL after waiting on a closed
socket, as that was an issue while fixing the previous code. A test for
a wait on a named pipe is also added, to ensure there's not regression
there.