Remove use of asm_routines_end from fdefn_callee_lispobj()
This was a source of error for space relocation because it matters
whether you mean the asm routines where they were supposed to be
or where they actually are.  Instead, an fdefn's "raw addr" slot
directly encodes both the jump target (using 5 bytes for the JMP
instruction), and the offset to the tagged object being jumped to.
The exception is that 0 offset means sb-fasl:*assembler-routines*.