Add unique serial# to code objects, 64-bit only
This will allow ordering code in a well-defined way, by creation order.
It may not be the optimal order from a locality sense,
but is better for build reproducibility than the current heuristic,
and we don't know that the heurstic improves anything to begin with.

Put ALLOCATE-CODE-OBJECT in 'alloc' where it seems generally happier.