Starting PPC64 port - add :PPC64 Lisp feature
This builds up to but not including second genesis.

You have to explicitly choose --arch=ppc64 to compile.

The C runtime does not compile if ldso-stubs.S is included in the link
step due to "unresolvable R_PPC64_REL24 against `{symbol}@@GLIBC_2.3'"
for every symbol, because we haven't figured out enough of the ABI yet.