Really fix printf conversion warnings. Really.
Change e98378ca00 fixed Windows warnings but caused Darwin warnings.  The reason this
is so messed up for us - and it needn't be - is that we let the backend specify each
type to use for "our" abstractions on top of system-defined types such as size_t,
and then we *also* try to say that formatting such a thing can be done using one of
the now-standardized macros specifying a printf format. Well, you can't have both.
If backend X wants to say that "our" size_t is actually uword_t versus size_t versus
'unsigned long', then inevitably it is impossible to format that thing without casting
it into the type that the printf format expects. So the bug is that we allow our
type to alias a random thing.  os_vm_size_t should just be uintptr_t everywhere.