Add SB-EXT:MAKE-WEAK-VECTOR and SB-EXT:WEAK-VECTOR-P
A weak vector is logically equivalent to a simple-vector of weak pointers
but requires less storage by removing a level of indirection.
Elements which are not strongly pointed to automatically become NIL
when the unreferenced objects are garbage-collected.

Such vectors are not more efficient for GC to handle versus a similarly-sized
vector of weak pointers, so an algorithm which uses a weak vector should
probably have some reason for wanting to use a vector per se.

The vector type is SIMPLE-VECTOR, not a strict subtype in the type-theoretic
sense, but you can inquire whether a vector is weak using WEAK-VECTOR-P.