x86-64: Use MOVQ, not MOVD, to move qwords.
Henceforth, MOVD will only move dwords. While technically incompatible,
this changes no behavior that anybody should care about. As an example of why
this is better, if you take %MAKE-SIMD-PACK-UB64 vop and hand-transpile to C:
    typedef long long2 __attribute__((ext_vector_type(2)));
    long2 foo(long a, long b) {
      long2 v;
      v.x = a; v.y = b;
      return v;
    }

and then compile that C to assembly with "cc -S -O -o foo.s foo.c" to get:
    movq %rsi, %xmm1
    movq %rdi, %xmm0
    punpcklqdq %xmm1, %xmm0

the issue is clear: a C 'long' is moved using MOVQ. But if it's just a name,
and our operation is the same either way, why change? The comments explain.