first trace file!
We need to emit labels that we have generated, otherwise the metadata
isn't consistent and the compiler croaks.

We also need EMIT-BLOCK-HEADER and EMIT-ERROR-BREAK functions for
emittage, ERROR-TRAP and related variables, and the %TEST-HEADERS
function for automatically-generated type tests.

With this, we can compile src/code/scratch.lisp "successfully", and
also successfully generate the trace file to prove it.  (Said
compilation, of course, produces no actual instructions, and the trace
file reveals some issues with our current setup, but it's our first
tangible output artifact.  Onward!)