Add another operation to editcore
'copy' mode embeds an entire Lisp core into an ELF section with no alteration.
The resulting fully linked ELF executable is self-contained and has the
desirable aspect that it allows further allocation to immobile code space.
The disadvantage is that it does not have ELF symbols for Lisp entry points.