x86-64: Add better abstraction for CPU register in the assembler
Any TN which refers to a register is converted to an instance of REG
just prior to calling the instruction emitter. This makes it simple
to distinguish registers from stack locations. Additionally the function
which returns a register in a different size no longer has anything
to do with the VM's storage classes.

We can henceforth remove all of the global variables which hold
TNs in sizes other than :qword (e.g. AL-TN, AX-TN, EAX-TN)
and we can probably remove their storage classes, except for potential
problems with third-party libraries that use define-vop and assume
that they can request vop temporaries in arbitrary sizes.