Fix build order and warn about duplicate DEFINE-SOURCE-TRANSFORM
The prior changes which removed features from make-config.sh, instead testing
for presence of vops to conditionalize some backend-independent code succumbed
to the problem noted in the change which added the :vop-translates reader
(change 3cd585ce03), namely, the build must be properly ordered.

As a result of mis-ordering, the ash-right transform (etc) was not built into
the cross-compiler, and the target code was missing some optimizations.

To rememdy this, 'src/compiler/srctran' must be compiled later, which exposed
an issue that we relied on overwriting a general source-transforms with one
provided by target/arith.