Remove junk from MAP-SEGMENT-INSTRUCTIONS
This function should not be responsible for special handling of bytes that
are not instructions.  Once upon a time it was necessary for the disassembler
to figure out how NOT to disassemble by accident (and mis-decode) raw data
following the last simple-fun or preceding the first simple-fun.

It's easier to just start disassembling at the first simple-fun, and doesn't
add significant value to show raw words considering that we can't render as
the necessarily correct data type anyway (like if it's complex-double-float).