Add identical-code-folding "link" step
ICF treats code objects as blobs of bytes, in the manner of string coalescing.
Two strategies are provided. The weak strategy will fold only if nothing points
to a code object other than at most an #<fdefn>, which gives some degree of
surety that nothing cares about function pointers per se. The aggressive
level removes that constraint and will rewrite pointers anywhere in the heap.
Either strategy assumes that function name, lambda list, docstring, type,
xrefs, and source form are irrelevant, and do not need to be compared.
Otherwise nothing could ever be folded, in as much as simple-funs can not
point to blobs of code elsewhere (though that may change in the future).

In the interest of caution, neither strategy is used by default, except as
the final step in self-build. The effects on our own build are fairly obvious.
There are some amusing but acceptable results, for example:
 #'SB-PCL::CLOS-SLOTS-REF becomes #<FUNCTION SVREF>
but
 #'(SETF SVREF) becomes #<FUNCTION (SETF SB-PCL::CLOS-SLOTS-REF).

Whereas for the baseline SBCL build, ICF achieves not much, in our real use
it removes 5MB of code and 2MB of debug infos. This is primarily from
out-of-line structure accessors in 0 safety. So if you have 1000 defstruct
forms, you have that many copies of (%INSTANCE-REF X n) for various N.

See also the paper referenced in a comment showing why ICF can be unsafe or
safe in C++ and what the linker tries to do about it. Any time the code takes
a function pointer, there might be a problem if it expects certain pointer
invariants. The C++ compiler can notice at the source level the "&" operator.
Absent that, there could not be a function pointer. But in Lisp, you can get
function pointers whenever you want, so ICF is never provably admissible.
All the same, the spec doesn't say that #'CAR and #'FIRST are non-EQ functions,
and it also doesn't say that COMPILE isn't memoizable.

We should probably make this a keyword argument to save-lisp-and-die.
Meanwhile, do with it as you will.