Put more burden on the Lisp code prior to immobile space defrag
Compute in Lisp the location of each operand that potentially needs fixing
as well as the referred-to Lisp object so that C doesn't do the latter.
In this manner, C code doesn't have to understand machine instruction formats
and doesn't have to be able to search for an object while defragmenting.