Prevent %EXIT-OTHER-THREADS from deadlocking with finalizer thread
Previously, %EXIT-OTHER-THREADS grabbed *MAKE-THREAD-LOCK* and tried
to join the finalizer thread afterwards. A garbage collection between
the two points could trigger RUN-PENDING-FINALIZERS and thus an
attempt to spawn the finalizer thread, leading to a deadlock.

To prevent this, %EXIT-OTHER-THREADS now calls FINALIZER-THREAD-STOP
before grabbing *MAKE-THREAD-LOCK*. Should a garbage collection occur
after grabbing the lock, finalizers are run synchronously (as they are
during later phases of the shutdown) avoiding the deadlock.