Fix weak hash-table culling bug
fullcgc did not update the PTEs after consuming conses, so depending on what
the allocator did, it might have gotten lucky by randomly NOT sweeping the
conses that it just made, but often it got unlucky and "correctly" swept the
memory, which, appearing unused, blasted the queue to smithereens.
This change ensures that the memory taken by those conses appears in-use.