Remove unnecessary code from ansify
Genesis works fine regardless of the home package of symbols exported from the
COMMON-LISP package. CLISP is known to work as a build host, and it has many
CLOS-related symbols homed in its CLOS package but exported from CL.

As a further test I built a "weird" SBCL with the following alterations:
  * (symbol-package 'FLOATING-POINT-INEXACT) => #<PACKAGE "SB-EXT">
  * (symbol-package 'DEBUG) => #<PACKAGE "SB-EXT">
as the comment says was the case with CMUCL. That build worked fine as host for
a normal SBCL with symbols ending up where they should. And as a final test,
an even weirder SBCL with every CL symbol homed in a random package also worked,
however it revealed a cross-compiler bug in STD-ATOM-TYPE-SPECIFIER-P.