Solve the problem of SB! embedded in format control strings
This is part 1 of a series of changes that will reduce the chance of
getting your sb! and sb- wrong in any context.

A literal format control string that is dumped by the cross-compiler
will have all ~/pkg:sym/ directives parsed out at xc-time and then dumped
as a string plus a list of symbols to stuff back in to that string.
The partially tokenized object also serves as a place to fully tokenize
and memoize, which is slightly faster than using the control-string cache.

Furthermore it makes packages renaming possible as shown:
* (sb-kernel:code-header-ref
   (sb-kernel:fun-code-header #'sb-alien::missing-alien-operation-error) 4)
=> #<SB-FORMAT::FMT-CONTROL "Cannot ~A aliens of type ~/SB-IMPL:PRINT-TYPE-SPECIFIER/.">

* (sb-ext:without-package-locks (rename-package "SB-IMPL" "FRUITY"))
=> #<PACKAGE "FRUITY">

* (sb-kernel:code-header-ref
   (sb-kernel:fun-code-header #'sb-alien::missing-alien-operation-error) 4)
=> #<SB-FORMAT::FMT-CONTROL "Cannot ~A aliens of type ~/FRUITY:PRINT-TYPE-SPECIFIER/.">