Exclude format-time from make-host-1 and change some wording
The word "written" printed twice per input file - once for the last modified
time of the source, and then again to say that we did something - always felt
odd to me.  So change the latter to "wrote" which is clearly an action verb.
But seriously, time formatting routines are not an essential piece of
the cross-compiler, so don't show timestamps in self build.