Strengthen the tests in verify_range
Checking for consistency of the page write-protect with nonexistence
of old->young pointers can/should be done for any managed space.
That test was inadvertently skipped for pointers to immobile space.
Additionally we should check for nonexistence of forwarding pointers
in immobile space, as they're used during defrag and must not persist.